// starter/packages/common/src/net/types.ts
export interface HttpError extends Error {
	DeveloperMessage: string;
	UserMessage?: string;
	ErrorCode?: number;
	MoreInfo?: string;
	StatusCode?: HttpStatusCode;
	OriginalError: AxiosError<any>;
}